+++ src/contexts/remoteConfigs.tsx
@@ -1,8 +1,8 @@
-import React, { createContext, useContext, useEffect, useState } from 'react';
-import api from '../services/api';
+import React, { createContext, useContext, useEffect, useState } from "react";
+import RemoteConfig from "@react-native-firebase/remote-config";
+import { minutesToMilliseconds } from "date-fns";
 
 interface RemoteConfigContextProps {
-  getRemoteConfigs: (type?: "premium" | "default" | "all") => Promise<Configs>,
   allConfigs: Configs;
   userConfigs: UserConfigs;
 }
@@ -25,70 +25,84 @@ interface UserConfigs {
   messageLength: number;
 }
 
-const RemoteConfigsContext = createContext({} as RemoteConfigContextProps)
+const RemoteConfigsContext = createContext({} as RemoteConfigContextProps);
 
-const RemoteConfigsProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-
-  const [allConfigs, setAllConfigs] = useState<Configs>({} as Configs)
-  const [userConfigs, setUserConfigs] = useState<UserConfigs>({} as UserConfigs)
-
+const RemoteConfigsProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  const [allConfigs, setAllConfigs] = useState<Configs>({} as Configs);
+  const [userConfigs, setUserConfigs] = useState<UserConfigs>(
+    {} as UserConfigs
+  );
 
   useEffect(() => {
     (async () => {
-      const configs = await getRemoteConfigs("all")      
-      setAllConfigs(configs)
-      
-      setInterval(async () => {
-        const configs = await getRemoteConfigs("all")      
-        setAllConfigs(configs)
-      }, 300000)
-    })()
-  }, [])
+      await RemoteConfig().setConfigSettings({
+        minimumFetchIntervalMillis: minutesToMilliseconds(30),
+      });
+      await RemoteConfig()
+        .setDefaults({
+          default_max_groups: 2,
+          premium_max_groups: 10,
+          default_file_upload: 12,
+          premium_file_upload: 120,
+          default_max_participants: 200,
+          premium_max_participants: 1000,
+          default_max_message_length: 500,
+          premium_max_message_length: 5000,
+        })
+        .then(() => RemoteConfig().fetchAndActivate())
+        .then((fetched) => {});
 
-  useEffect(() => {
-    updateUserConfigs()
-  }, [allConfigs])
-
-  const getRemoteConfigs = async (type?: string) => {
-    const res = await api.get(`/app/configs?type=${type || "all"}`)
-    return res.data as Configs
-  }
+      await updateUserConfigs();
+    })();
+  }, []);
 
   const updateUserConfigs = async () => {
+    const values = RemoteConfig().getAll();
+    const configs = {} as Configs;
+
+    Object.entries(values).forEach(($) => {
+      const [key, entry] = $;
+
+      allConfigs[key] = entry.asString();
+    });
+
+    setAllConfigs(configs)
+
     if (false) {
       return setUserConfigs({
-        fileUpload: parseFloat(allConfigs.premium_file_upload),
-        amountGroups: parseInt(allConfigs.premium_max_groups),
-        amountParticipants: parseInt(allConfigs.premium_max_participants),
-        messageLength: parseInt(allConfigs.premium_max_message_length)
-      })
+        fileUpload: parseFloat(configs.premium_file_upload),
+        amountGroups: parseInt(configs.premium_max_groups),
+        amountParticipants: parseInt(configs.premium_max_participants),
+        messageLength: parseInt(configs.premium_max_message_length),
+      });
     }
 
     return setUserConfigs({
-      fileUpload: parseFloat(allConfigs.default_file_upload),
-      amountGroups: parseInt(allConfigs.default_max_groups),
-      amountParticipants: parseInt(allConfigs.default_max_participants),
-      messageLength: parseInt(allConfigs.default_max_message_length)
-    })
-  }
+      fileUpload: Number(configs.premium_file_upload),
+      amountGroups: Number(configs.default_max_groups),
+      amountParticipants: Number(configs.default_max_participants),
+      messageLength: Number(configs.default_max_message_length),
+    });
+  };
 
   return (
     <RemoteConfigsContext.Provider
       value={{
-        getRemoteConfigs,
         allConfigs,
-        userConfigs
+        userConfigs,
       }}
     >
       {children}
     </RemoteConfigsContext.Provider>
   );
-}
+};
 
 const useRemoteConfigs = () => {
-  const configsContext = useContext(RemoteConfigsContext)
+  const configsContext = useContext(RemoteConfigsContext);
 
-  return configsContext
-}
+  return configsContext;
+};
 
 export { RemoteConfigsProvider, useRemoteConfigs };